Precisely define the width of most integer values

+++ GitarooFileReader/CHC/CHC.h
@@ -45,8 +45,8 @@ private:
 	// Hold data for all 8 audio channels
 	struct AudioChannel
 	{
-		unsigned long volume = 32767;
-		unsigned long pan = 16383;
+		uint32_t volume = 32767;
+		uint32_t pan = 16383;
 	} m_audio[8];
 
 	// Gameplay speed
@@ -127,7 +127,7 @@ public:
 	enum class Phase { INTRO, CHARGE, BATTLE, FINAL_AG, HARMONY, END, FINAL_I };
 private:
 	//Index pulled from the list of SSQs in the CHC
-	unsigned long m_index = 0;
+	uint32_t m_index = 0;
 	// Section name
 	// 16 characters + a null character
 	char m_name[17] = { 0 };
@@ -139,9 +139,9 @@ private:
 	//SSQ frame range
 	SSQRange m_frames;
 	// Saves whether the subsections are correctly ordered
-	unsigned long m_organized = false;
+	uint32_t m_organized = false;
 	// Value that holds how the section is swapped
-	unsigned long m_swapped = 0;
+	uint32_t m_swapped = 0;
 	// Just junk, saved for consistency
 	char m_junk[16] = { 0 };
 
@@ -150,14 +150,14 @@ private:
 	// BPM for the section
 	float m_tempo = 0;
 	// Total duration in samples
-	unsigned long m_duration = 0;
+	uint32_t m_duration = 0;
 	// Determines what actions to take after the section
 	struct Condition
 	{
-		unsigned long m_type;
+		uint32_t m_type;
 		float m_argument;
-		long m_trueEffect;
-		long m_falseEffect;
+		int32_t m_trueEffect;
+		int32_t m_falseEffect;
 		Condition();
 		Condition(FILE* inFile);
 		Condition(const Condition& cond) = default;
@@ -166,15 +166,15 @@ private:
 	// Vector of all conditions
 	std::vector<Condition> m_conditions;
 	// Number of players assigned to the section, always 4
-	unsigned long m_numPlayers = 4;
+	uint32_t m_numPlayers = 4;
 	// Number of charts/subsections per player
-	unsigned long m_numCharts = 1;
+	uint32_t m_numCharts = 1;
 	// Parent CHC
 	CHC* m_parent = nullptr;
 public:
 	// Vector of all charts/subsections
 	std::vector<Chart> m_charts;
-	static const long s_SAMPLE_GAP = 1800;
+	static const int32_t s_SAMPLE_GAP = 1800;
 	SongSection(CHC* parent);
 	SongSection();
 	SongSection(CHC* parent, FILE* inFile);
@@ -186,7 +186,7 @@ public:
 
 	void create(FILE* outFile);
 	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile);
-	unsigned long getIndex() const { return m_index; }
+	uint32_t getIndex() const { return m_index; }
 	// Returns name C-string (size: 16)
 	char* getName() { return m_name; }
 	// Returns audio C-string (size: 16)
@@ -194,20 +194,20 @@ public:
 	// Returns whether the section is organized
 	bool getOrganized() const { return m_organized; }
 	// Sets organized to the provided value
-	void setOrganized(unsigned long org);
+	void setOrganized(uint32_t org);
 	// Returns swap value
-	unsigned long getSwapped() const { return m_swapped; }
+	uint32_t getSwapped() const { return m_swapped; }
 	// Sets swap value
 	void setSwapped(char swap) { m_swapped = swap; }
 	// Returns the byte size of the section
-	unsigned long getSize() const;
+	uint32_t getSize() const;
 	// Returns a section's phase type
 	Phase getPhase() const { return m_battlePhase; }
 	// Sets the section's phase type to the provided value
 	void setPhase(Phase ph) { m_battlePhase = ph; }
 	// Sets the section's phase type to the provided value converted into a phase type
 	// If the provided value is 6 or greater, battlePhase is set to FINAL_I
-	void setPhase(unsigned long ph)
+	void setPhase(uint32_t ph)
 	{
 		if (ph < 6)
 			m_battlePhase = static_cast<SongSection::Phase>(ph);
@@ -217,9 +217,9 @@ public:
 	// Returns the tempo for the section
 	float getTempo() const { return m_tempo; }
 	// Returns the duration of the section
-	unsigned long getDuration() const { return m_duration; }
+	uint32_t getDuration() const { return m_duration; }
 	// Sets the duration of the section to the provided value
-	void setDuration(unsigned long dur) { m_duration = dur; }
+	void setDuration(uint32_t dur) { m_duration = dur; }
 	// Returns the number of conditions in the section
 	size_t getNumCondtions() const { return m_conditions.size(); }
 	// Adds a new condition to the end of the section's condition list
@@ -234,9 +234,9 @@ public:
 	bool removeCondition(size_t);
 	// Returns the num of players assigned to this section.
 	// Will usually be 4
-	unsigned long getNumPlayers() const { return m_numPlayers; }
+	uint32_t getNumPlayers() const { return m_numPlayers; }
 	// Returns the total number of the charts/subsections in this section
-	unsigned long getNumCharts() const { return m_numCharts; }
+	uint32_t getNumCharts() const { return m_numCharts; }
 	void clearConditions();
 	void operator++();
 	bool operator--();
@@ -270,10 +270,10 @@ private:
 	char m_junk[16] = { 0 };
 
 	// 'Center' point that all notes revolve around
-	long m_pivotTime = 0;
+	int32_t m_pivotTime = 0;
 	// Optional value noting when to transition to another chart/subsection
 	// Unused in the game
-	long m_endTime = 0;
+	int32_t m_endTime = 0;
 public:
 	// Vector of all trace lines
 	std::vector<Traceline> m_tracelines;
@@ -290,22 +290,22 @@ public:
 	void create(FILE* outFile);
 	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile);
 	// Returns the byte size of the chart/subsection
-	unsigned long getSize() const;
+	uint32_t getSize() const;
 	// Returns junk C-string (size: 16)
 	char* getJunk() { return m_junk; }
 	// Copies C-string newJunk to junk
 	// Size "count" will be limited to a max of 16
 	void setJunk(char* newJunk, rsize_t count = 16);
 	// Returns the pivot time for the chart/subsection
-	long getPivotTime() const { return m_pivotTime; }
+	int32_t getPivotTime() const { return m_pivotTime; }
 	// Sets the pivot time for the chart/subsection to the provided value
-	void setPivotTime(long piv) { m_pivotTime = piv; }
+	void setPivotTime(int32_t piv) { m_pivotTime = piv; }
 	// Adjusts the pivot time for the chart/subsection by the provided value
-	void adjustPivotTime(long difference) { m_pivotTime += difference; }
+	void adjustPivotTime(int32_t difference) { m_pivotTime += difference; }
 	// Returns the end time for the chart/subsection
-	long getEndTime() const { return m_endTime; }
+	int32_t getEndTime() const { return m_endTime; }
 	// Sets the end time for the chart/subsection to the provided value
-	void setEndTime(long piv) { m_endTime = piv; }
+	void setEndTime(int32_t piv) { m_endTime = piv; }
 	// Returns the number of the trace lines in the chart/subsection
 	size_t getNumTracelines() const { return m_tracelines.size(); }
 	template<class... Args>
@@ -347,15 +347,15 @@ class Note
 	friend class CHC;
 public:
 	// Displacement from chart pivot time in samples
-	long m_pivotAlpha;
+	int32_t m_pivotAlpha;
 	Note();
-	Note(long alpha);
+	Note(int32_t alpha);
 	Note(FILE* inFile);
 	Note(const Note& note) = default;
 	virtual Note& operator=(const Note& note);
 	virtual void create(FILE* outFile);
 	// Adjusts the note's pivot alpha by the provided value
-	void adjustPivotAlpha(long change) { m_pivotAlpha += change; }
+	void adjustPivotAlpha(int32_t change) { m_pivotAlpha += change; }
 	bool operator==(const Note& other) const { return m_pivotAlpha == other.m_pivotAlpha; }
 	auto operator<=>(const Note& other) const { return m_pivotAlpha <=> other.m_pivotAlpha; }
 };
@@ -365,9 +365,9 @@ class Path : public Note
 	friend class CHC;
 public:
 	// Duration of the note in samples
-	unsigned long m_duration;
+	uint32_t m_duration;
 	Path();
-	Path(long alpha, unsigned long dur = 1);
+	Path(int32_t alpha, uint32_t dur = 1);
 	Path(FILE* inFile);
 	Path(const Note& note);
 	Note& operator=(const Note& note);
@@ -375,23 +375,23 @@ public:
 	// Adjusts the path note's duration to the provided value
 	// Returns whether the resulting duration is >= 1
 	// If not, duration gets set to 1 and false is returned
-	bool adjustDuration(long change);
+	bool adjustDuration(int32_t change);
 	// Returns the result of adding the note's pivot alpha with its duration
-	long getEndAlpha() const { return m_pivotAlpha + m_duration; }
+	int32_t getEndAlpha() const { return m_pivotAlpha + m_duration; }
 	// Attempts to set the path note's pivot alpha to the provided value.
 	// Will stretch or shrink the note's duration to maintain the same end alpha.
 	// If the result would be a duration < 1, then no changes are made and false is returned.
 	// Otherwise, return true.
-	bool changePivotAlpha(const long alpha);
+	bool changePivotAlpha(const int32_t alpha);
 	// Attempts to set the path note's end alpha to the provided value.
 	// Automatically adjusts duration to match.
 	// If the result would be a duration < 1, then no changes are made and false is returned.
 	// Otherwise, return true.
-	bool changeEndAlpha(const long endAlpha);
+	bool changeEndAlpha(const int32_t endAlpha);
 	// Returns whether the provided alpha value lands inside the path note's range of values
-	bool contains(const long alpha) const { return alpha >= m_pivotAlpha && alpha < m_pivotAlpha + (long)m_duration; }
+	bool contains(const int32_t alpha) const { return alpha >= m_pivotAlpha && alpha < m_pivotAlpha + (int32_t)m_duration; }
 	// Returns whether the pivot alpha of the provided note lands inside the path note's range of values
-	bool contains(Note& note) const { return note.m_pivotAlpha >= m_pivotAlpha && note.m_pivotAlpha < m_pivotAlpha + (long)m_duration; }
+	bool contains(Note& note) const { return note.m_pivotAlpha >= m_pivotAlpha && note.m_pivotAlpha < m_pivotAlpha + (int32_t)m_duration; }
 };
 
 class Traceline : public Path
@@ -402,15 +402,15 @@ public:
 	// Held in Radians
 	float m_angle;
 	// Holds whether the entire trace line is curved
-	unsigned long m_curve;
+	uint32_t m_curve;
 	Traceline();
-	Traceline(long alpha, unsigned long dur = 1, float ang = 0, unsigned long cur = false);
+	Traceline(int32_t alpha, uint32_t dur = 1, float ang = 0, uint32_t cur = false);
 	Traceline(FILE* inFile);
 	Traceline(const Note& note);
 	Note& operator=(const Note& note);
 
 	void create(FILE* outFile);
-	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile, const long pivotTime);
+	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile, const int32_t pivotTime);
 	// Adjusts the radian angle value of the trace line by the provided value
 	void adJustAngle(float change) { m_angle += change; }
 };
@@ -424,30 +424,30 @@ private:
 	char m_junk[12] = { 0 };
 	// Color value used in Clone Hero & color patch exporting
 	// 0 = not used
-	unsigned long m_color;
+	uint32_t m_color;
 public:
 	// Boolean for whether the phrase bar has a start cap
-	unsigned long m_start;
+	uint32_t m_start;
 	// Boolean for whether the phrase bar has an end cap
-	unsigned long m_end;
+	uint32_t m_end;
 	// Which character animation to play when hit (if "start" is true)
-	unsigned long m_animation;
+	uint32_t m_animation;
 	Phrase();
-	Phrase(long alpha, unsigned long dur = 1, unsigned long start = true, unsigned long end = true, unsigned long anim = 0, unsigned long color = 0);
+	Phrase(int32_t alpha, uint32_t dur = 1, uint32_t start = true, uint32_t end = true, uint32_t anim = 0, uint32_t color = 0);
 	Phrase(FILE* inFile);
 	Phrase(const Note& note);
 	Note& operator=(const Note& note);
 
 	void create(FILE* outFile);
-	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile, const long pivotTime);
+	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile, const int32_t pivotTime);
 	// Returns the color value of the phrase bar
-	unsigned long getColor() const { return m_color; }
+	uint32_t getColor() const { return m_color; }
 	// Sets the color value of the phrase bar
-	void setColor(unsigned long col) { m_color = col; }
+	void setColor(uint32_t col) { m_color = col; }
 	// Sets the color value of the phrase bar
-	void addColor(unsigned long col) { m_color |= col; }
+	void addColor(uint32_t col) { m_color |= col; }
 	// Sets the color value of the phrase bar
-	void removeColor(unsigned long col) { m_color &= ~col; }
+	void removeColor(uint32_t col) { m_color &= ~col; }
 	// Returns junk C-string (size: 12)
 	char* getJunk() { return m_junk; }
 };
@@ -461,13 +461,13 @@ public:
 	// 1 - X
 	// 2 - O
 	// 3 - /\.
-	unsigned long m_button = 0;
+	uint32_t m_button = 0;
 	Guard();
-	Guard(long alpha, unsigned long but = 0);
+	Guard(int32_t alpha, uint32_t but = 0);
 	Guard(FILE* inFile);
 	Guard(const Note& note);
 	Note& operator=(const Note& note);
 
 	void create(FILE* outFile);
-	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile, const long pivotTime);
+	void write_to_txt(FILE*& txtFile, FILE*& simpleTxtFile, const int32_t pivotTime);
 };
